refactor(checker): split duplicate identifier helpers
